+++ ansible_deployment/inventory_plugins/inventory_plugin.py
@@ -1,6 +1,7 @@
 """
 Inventory class skeleton.
 """
+from ansible_deployment import SSHKeypair
 
 
 class InventoryPlugin:
@@ -35,6 +36,7 @@ class InventoryPlugin:
         }
         self.host_vars = {}
         self.group_vars = {}
+        self.ssh_keypair = SSHKeypair()
         self.group_vars["all"] = {"ansible_user": "ansible",
                                   "ansible_ssh_private_key_file": "./.ssh/id_rsa",
                                   "ansible_ssh_public_key_file": "./.ssh/id_rsa.pub"}